#380: Consider using an efficient reader.
