+++ ladder/ladder-parser/src/main/scala/com/actionfps/ladder/parser/TsvExtractEfficient.scala
@@ -1,0 +1,12 @@
+package com.actionfps.ladder.parser
+
+import java.nio.file.Path
+
+/**
+  * Created by william on 13/5/17.
+  */
+object TsvExtractEfficient {
+  def buildAggregateEfficient(path: Path, nickToUser: NickToUser): Aggregate = {
+    ???
+  }
+}
